Quite a few changes in this commit!
Films.php - updated to bring up a page that updates depending on which film you clicked, rather than having a page for each film individually.

Member.php - added two new function borrowedmem() prints a user's first name and "you have borrowed; dueDate() just prints "Your due date is" and then populates the duedate from the database in another file.

fm.php - the .php file that loads when you click on a film. Queries the database and pulls all information from it in regards to the film that has been clicked on.

loan.php - the page that appears when you click the "Loan Film" button. On click it inserts a row into the onloan page, updates the Films table with the availability and loan count and opens a new page with the functions from Members class about what film they loaned and when it's due. Not very pretty atm, but does the trick.

Images - slightly update the paths. Will need to add code for the images table that'll store all image paths.