Fm.php - checks the users age against the film rating. If the user is too young to loan it out it won’t display the “Loan Film” button. However, to make this work, I had to change the U to 0 and PG to 12 in the database.
MembersAccount.php - added a Return Loan button. When clicked, it updates the film’s due date to today’s date.  Would be good if the “Your Loan History” was actually pulling from the database to fully test this, but I assume we’ll need sessions for that before anything. So that then we could set it to GET user_ID from sessions and fm_ID from the “Your Loan History” table. At the moment, they’re just manually put in.

ReturnLoan.php - code that is executed when a user returns their loan.